Address gcc warning: comparison of integer expressions of different signedness

+++ src/sbbs3/netmail.cpp
@@ -375,7 +375,7 @@ void sbbs_t::qwktonetmail(FILE *rep, char *block, char *into, uchar fromhub)
 		return; 
 	}
 	memcpy((char *)qwkbuf,block,QWK_BLOCK_LEN);
-	if(fread(qwkbuf+QWK_BLOCK_LEN, QWK_BLOCK_LEN, n-1, rep) != n-1) {
+	if(fread(qwkbuf+QWK_BLOCK_LEN, QWK_BLOCK_LEN, n-1, rep) != (size_t)n-1) {
 		errormsg(WHERE, ERR_READ, "QWK block", n-1);
 		return;
 	}
